Rename SelectedBlockAttributes to BlockXray
